fix incorrect reference to serialiser
I'm trying to debug an error that only occurs with the TastyPieAPI and only when I'm using drest from my server. 

Although this may not be related to my problem, it does seem to be a minor mistake. (although I realise at the moment it doesn't matter since the base class already has its meta set to the same serialiser)